requirements.txt
lightgbm>=4.5,<5
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.5,<2
matplotlib>=3.8,<4
streamlit>=1.41,<2
pydantic>=2.10,<3
python-dotenv>=1.0,<2
jinja2>=3.1,<4
pytest>=8.3,<9
